Move shared form button to subdirectory for clarity
