+++ apps/website/src/components/admin/bingo/BingoForm.tsx
@@ -24,7 +24,7 @@ import {
 
 import { type BingoSchema } from "@/server/db/bingos";
 
-import { Button, defaultButtonClasses } from "@/components/shared/Button";
+import { Button, defaultButtonClasses } from "@/components/shared/form/Button";
 import { TextField } from "@/components/shared/form/TextField";
 import { FieldGroup } from "@/components/shared/form/FieldGroup";
 import { Fieldset } from "@/components/shared/form/Fieldset";
