+++ apps/website/src/components/admin/notifications/SendNotificationForm.tsx
@@ -16,7 +16,7 @@ import IconLoading from "@/icons/IconLoading";
 
 import { TextField } from "@/components/shared/form/TextField";
 import { TextAreaField } from "@/components/shared/form/TextAreaField";
-import { Button, defaultButtonClasses } from "@/components/shared/Button";
+import { Button, defaultButtonClasses } from "@/components/shared/form/Button";
 import { SelectBoxField } from "@/components/shared/form/SelectBoxField";
 import {
   UploadAttachmentsField,
