+++ apps/website/src/components/admin/short-links/ShortLinks.tsx
@@ -6,7 +6,7 @@ import {
   Button,
   dangerButtonClasses,
   LinkButton,
-} from "@/components/shared/Button";
+} from "@/components/shared/form/Button";
 import { ModalDialog } from "@/components/shared/ModalDialog";
 import { Headline } from "@/components/admin/Headline";
 import { Panel } from "@/components/admin/Panel";
