+++ apps/website/src/components/admin/show-and-tell/AdminShowAndTellEntriesPanel.tsx
@@ -1,7 +1,7 @@
 import { Fragment, useCallback } from "react";
 import type { inferRouterOutputs } from "@trpc/server";
 import { trpc } from "@/utils/trpc";
-import { Button } from "@/components/shared/Button";
+import { Button } from "@/components/shared/form/Button";
 import IconLoading from "@/icons/IconLoading";
 import type { AppRouter } from "@/server/trpc/router/_app";
 import { Panel } from "../Panel";
