+++ apps/website/src/components/admin/twitch/TwitchChannelForm.tsx
@@ -6,7 +6,7 @@ import { useRouter } from "next/router";
 
 import { trpc } from "@/utils/trpc";
 
-import { Button, defaultButtonClasses } from "@/components/shared/Button";
+import { Button, defaultButtonClasses } from "@/components/shared/form/Button";
 import { TextField } from "@/components/shared/form/TextField";
 import { Fieldset } from "@/components/shared/form/Fieldset";
 import { MessageBox } from "@/components/shared/MessageBox";
