+++ apps/website/src/components/shared/AddEventButton.tsx
@@ -7,7 +7,7 @@ import IconCalendar from "@/icons/IconCalendar";
 import IconGoogleCalendar from "@/icons/IconGoogleCalendar";
 import IconOutlook from "@/icons/IconOutlook";
 
-import { Button } from "@/components/shared/Button";
+import { Button } from "@/components/shared/form/Button";
 import { PopoverButton } from "@/components/shared/PopoverButton";
 
 const ICS_FEED_NAME = "Alveus announcement";
