+++ apps/website/src/components/shared/PopoverButton.tsx
@@ -1,6 +1,6 @@
 import { Fragment, type MouseEventHandler, type ReactNode } from "react";
 import { Popover } from "@headlessui/react";
-import { Button, defaultButtonClasses } from "@/components/shared/Button";
+import { Button, defaultButtonClasses } from "@/components/shared/form/Button";
 
 export function PopoverButton({
   children,
