+++ apps/website/src/components/shared/form/ImageUploadAttachment.tsx
@@ -4,7 +4,7 @@ import { classes } from "@/utils/classes";
 
 import IconLoading from "@/icons/IconLoading";
 
-import { Button } from "@/components/shared/Button";
+import { Button } from "@/components/shared/form/Button";
 import type { FileReference } from "@/components/shared/form/UploadAttachmentsField";
 import IconTrash from "@/icons/IconTrash";
 import IconWarningTriangle from "@/icons/IconWarningTriangle";
